Use var galera_cluster_members to identify cluster
Create a new variable called galera_cluster_members which is a list of
the hosts that make up the galera cluster. Set the default value to
groups['galera_all'] which is the group used by openstack-ansible to
refer to the members of the cluster. Replace all other references to
groups['galera_all'] with the new variable. This allows other consumers
of this role to use their own group naming scheme.

Add a new task to verify that current host is a member of
galera_cluster_members.

Add the var galera_client_drop_config_file=true to the role dependency
galera_client, this is required because the default value for that var
in the galera_client role is based on membership of the galera_all
group.

Modify the functional testing to override galera_cluster_members, this
aims to prevent the return of any hard-coded role references.

Change-Id: I59af07217114a001cbebaa95a651919d53c9ec21