Add retry logic to mysql_upgrade
This change adds a retry to the mysql_upgrade command when performing
an upgrade. This is required in circumstances when the wsrep service
is not ready for application use. This is a common issue when restarting
the mysql service after package installation.

Example Error:
  Got error: 1047: WSREP has not yet prepared node for application
  use when selecting the database

Change-Id: Icd859531b6a6f1e04ecd1214eb7c401fa0fe8048
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>
(cherry picked from commit 8245978ea7acda108b860e8adf64407d0c6e17cf)