Fix handlers
The handler to restart mysql is using an incorrect registered variable
to check if the restart succeeded, leading to multiple service restarts
each time its notified.

The 'Remove stale .sst' and 'Restart mysql fall back' handlers will only
be notified if 'Restart mysql' is 'changed', but their when clauses
were checking for 'failed'. Regardless, their when clauses can be
removed.

Change-Id: Icc038d17b560b24bb9bf5fdd523835102ca97030