ensure mysql defaults file is sourced
The mysql initscript that comes with mariadb-10.0 does not source the
/etc/default/mysql file, meaning that the MYSQLD_STARTUP_TIMEOUT that
we set in there is being ignored.  This has the effect of causing some
nodes to timeout their startup too early, interrupting the initial SST,
and causing the cluster join to fail.

This commit adds a line to the initscript to ensure that we are
sourcing /etc/default/mysql correctly.

Closes-Bug: #1569446
Co-Authored-By: Jimmy McCrory <jimmy.mccrory@gmail.com>

Change-Id: I6af1cc660c2bbc2a49707cfab995659dc0167f2f
(cherry picked from commit b9a7361b28a7acb2dd05ef960cf474dc91486a6a)