Move systemd handler above restart handler
Since handlers are executed in the order in which they are defined
the handler to restart systemd should come before the handlers to
restart mysql. This avoids the error case where the playbook is run
multiple times and systemd refuses to restart mysql giving the
following message:

"Warning: mysql.service changed on disk. Run 'systemctl
daemon-reload' to reload units."

Closes-Bug: #1633472

Change-Id: I3962c0e8f9f1cf3c01a3160be4c744e9fbd1ecfe