[DOCS] Added release-name as a watermark to Docs.
Added a water mark to the role documentation
to make clear to deployers which release they're
working with.

Modified conf.py that adds watermark to the documentation according
to particular branch. If the current branch is master than
watermark is 'Pre-release' and if the current branch is
stable/<release-name> than watermark is <release-name>.

This is a combined implementation based on the following
patches:
- https://review.openstack.org/372565
- https://review.openstack.org/382637
- https://review.openstack.org/384140

Change-Id: Icc1c4f2c26df08aefcf6cc613232e650162e665e