Fix MYSQLD_STARTUP_TIMEOUT for systemd
The /etc/default/mysql file needs to be sourced for systemd to start
mysql with the appropriate MYSQLD_STARTUP_TIMEOUT value.

We are already adding the line to source the /etc/default/mysql file in
the init.d/mysql script, but we were skipping dropping the defaults file
when using "systemd".

Additionally, this groups the init file change and the default file
templating to be sequential.

Change-Id: If5dc328e67bde9d84f40c2b192c217e82e1987eb
Closes-Bug: #1634562
(cherry picked from commit 9a8fced0f214734f2de28e8178e06a04ce2235f7)