Allow a single-node MariaDB cluster to restart properly
With the implementation of https://review.openstack.org/382683 a single
MariaDB node has no peers configured, so there's no need to bootstrap
the cluster on restart.

This patch removes the condition in the handler which previously was
needed to handle the re-bootstrap during a single node cluster service
restart.

Closes-Bug: #1595143
Closes-Bug: #1639900
Related-Bug: #1624327
Change-Id: I599bbf0efa4e3d5abdf6d95c95d7983c464b3ae5