Don't install client packages via the galera_client role
The galera_server and galera_client roles both configure the apt or
yum repository for galera and percona. This can cause a conflict if
the version of the client and server differ.

This patch sets the galera_client dependency to deploy the /root/.my.cnf
file rather than installing all the client packages as well. The client
package is already deployed as part of the galera_server role
deployment.

Change-Id: I9534a02d4ce827a2f0fdf415a11e37808be941e5
Depends-On: I00d662a8afc7ddd4778787d31dc394a0ea3b1401