Remove percona repo if not used
If the ``use_percona_upstream`` value is changed from True to False,
don't remove the existing repository. This is also the source of a bug
in Mitaka: We were installing the repo by default and installed the
percona-xtrabackup tool with a deb file. If another playbook/role
installed percona-xtrabackup, it would override (and probably upgrade)
the deb installed version, which is undesired.

Combined backport of:
- https://review.openstack.org/#/c/417362/
- https://review.openstack.org/#/c/417776/
- https://review.openstack.org/#/c/416234/

Change-Id: Ibb2f01e6b61774a2517feb5baf4a5452b7f68a84
Closes-Bug: 1653947
Signed-off-by: Jean-Philippe Evrard <jean-philippe.evrard@rackspace.co.uk>