Use local facts for ansible_architecture
This patch ensure the local host's ansible facts are used for the
"ansible_architecture" fact, rather than the "physical_host". Since we
don't explicitly collect facts for "physical_host", and the
ansible_architecture fact is set for the local host container, we
don't need to try to retrieve the "physical_host" facts.

A new var "galera_package_arch" is created so that the architecture can
be overriden if it isn't being calculated correctly. Additionally, the
duplicate vars that are not specific to distributions have been moved
into defaults, and out of the specific distribution vars files.

Manual back-port of I5124c061faf124388a88eefee93d96e9a9e90d8f

Change-Id: I41494c2a1e6296a205f7cd5a5ee52b0aa90563aa
Closes-Bug: 1650384