Upgrade: Make removal list distribution-specific
When trying to execute an upgrade the removal fails
if packages in the list are included that cannot be
found in one of the repositories configured.

This patch endeavours to only remove packages that
will be found.

Change-Id: I67cc225b25fca005cdba3bf67793b4f23d6ae82a
(cherry picked from commit f939bf150fae4fe4a908d5a742396f04d4a7b471)