vars: main: Split percona packages from the galera server ones.
The role uses the 'use_percona_upstream' variable to determine if
specific external packages are going to be used instead of the distro or
upstream ones. The {qpress,percona}_package_path variables only make sense
if we are going to use such external packages so we convert their static
declaration into a 'set_fact' task. This also allows distributions to
override the default package list with an empty list if they do not plan
to use local percona packages.

Change-Id: Icb65f6c8c1654e24f73e1e5310516cd9b3568aad