tasks: Set Restart option for systemd service file
In order to have a consistent systemd behavior across distributions, we
define the Restart condition for the galera service file. The 'on-abort'
option is currently the default in the upstream service file and it's
also the most conservative one so we chose that to be the default option
in the Ansible role as well. We also set a reasonably low RestartSec
value as well.

Change-Id: If442802f7aff4473cbb573ba1c123cfc46bde9b3