Separate greenfield and upgrade tests
In order to independently test a greenfield deploy
and an upgrade, with the upgrade's initial deploy
using the previous stable branch role, we seperate
the two tests.

This makes it a more accurate test without having
to do funky workarounds in the role required to
cater for two versions of galera.

Depends-On: Ic881bf3acf6fb6daf71611c26322f02fb8b0b9be
Depends-On: I5c88b2019c01b44193a5d0df9299ecce6de52f01
Depends-On: I8b09c2520c84bb43353b4c56bac157259f71b041
Needed-By: I2aca4721e4e023d315f3fcdae0f931dc960ef2be
Change-Id: I0c74e511cec01ed51101b873784c4a90a7302da6