Fix MariaDB apt package pinning
The variables in vars are only loaded when the include_vars
task is run, which is after meta-dependencies are processed.

This patch ensures that the pinning is properly applied when
the meta-dependent role is executed.

Closes-Bug: #1712315
Change-Id: Ibc7766e96e478250b5e7d4788bb56dace974d25f
(cherry picked from commit c24d590d8acf808134e846f36a84fb23e88bd4ef)