Further simplify the upgrade check
The upgrade check does not need to see if the node is running. If the on
disk version does not match what we expect and the "galera_upgrade" flag
has not been set simply, instruct the run to fail. If galera_upgrade has
been set to "true" prep for the upgrade normally.

Change-Id: I943ea4dc8a2189dd766ff026cb44a00b9dd0f605
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>