Update default paths of SSL cert files
The default paths of SSL cert files has been changed to '/etc/ssl/certs'
to align with most other OpenStack-Ansible roles, and to match the
default location the the galera_client checks for an existing cert
distributing them to clients.

The default path for private key files remains '/etc/mysql/ssl'.

If any of these files changes it should trigger a restart of mysql.

Change-Id: Ifaa820e703110d7363095cf745ee3076fa263758