RedHat: Ensure that conflicting MariaDB packages are removed
When the galera_server role is deployed along with other roles, some
of them may pull mariadb packages as a dependency for the services
they deploy. However, this role uses MariaDB packages from the upstream
repository which conflict with those provided by the EPEL or the RDO
ones so we need to ensure that they are gone before we install the
upstream ones.

Change-Id: Ie800c1f4628eff24954f392fe6318206adfabddb
Closes-Bug: #1742206