Use a venv with previous ansible version for upgrades
When doing the role upgrade, we run every playbook with current's
branch ansible. This is normal for the infrastructure bits (which
are running on current branch), but not for the installation of
the previous branch's role, which should use previous branch's
environment.

Related-Bug: #1741471
Depends-On: I7ebb045885dd645b820de2b7f75b46c755c015f6
Change-Id: I7bc4dd644c48b403e9e669cb0ea460d9d1149410
(cherry picked from commit e4456ed4c1f7e3506e4698e17dc94b4778517bf9)