vars: Install qpress only when galera_xtrabackup_compression is true
The qpress library is only needed for xtrabackup compression. The
compression only occurs when a new node is joining the cluster
which shouldn't happen very often. The variable is set to true only
when the architecture is x86_64 since that's the only arch that's
supported by the upstream repository and CentOS and Ubuntu do not
provide qpress in the downstream distributions at all.

Change-Id: Iae8b5d0e0998c85ee453efc6620940c80243540e