Fix cache update after initial apt_repository fail
If apt_repository fails to update the apt cache after updating the
configuration, retries don't register there was a change and so no
attempt is made to update the cache by the module on the second attempt.
This failure can result in a failure to install packages.

This change puts responsibility for updating the cache on the apt module
task, this means updating the cache will get retried on failure and no
longer fail silently.

The changes made include:
- the two apt_repository tasks have their update_cache parameter set to
  no so it is clear where any update happens and there is no unnecessary
  duplication of cache updates.
- the use of retries is removed given the tasks no longer need to make
  use of the network.
- the apt_repository task result variables have been given unique names
  to ensure that the cache is updated if only the galera repository is
  added.

Change-Id: I91b7929113f720bc3c55343c481096657376302f
Partial-bug: 1750656