Start MySQL on localhost for galera_secure_node
When running MySQL on bare metal, it starts up the secure script
before configuring anything which means that it listens on all
interfaces by default.

In scenarios where the load balancer might be colocated with the
database server, it would cause a conflict and it would refuse
to start up (even if the database server is already configured
not to listen on that interface, the configuration only gets
dropped later in the play).

This sets the environment configuration to start it up on localhost
only and then resets it once it is done.  It should not affect
existing deployment but help with bare metal ones.

Change-Id: I883e32678ea3cc2eea09d23db5eecfce35aee6f2