Remove perl library from Ubuntu installs
It seems that the package was added a long time ago [1], and might
have been needed for the percona tooling. This makes sense if we're
downloading the dep - but we're no longer doing that and can rely
on apt's resolver to find the correct packages.

[1] https://review.openstack.org/238748

Change-Id: I742c4ef9a201384787d5a9aeafac4ff7b0bdc353
(cherry picked from commit 6f1d10063aac57f66dd7549028e090aed00dca7a)