Correct cluster name check
When checking for failed states the cluster name is gathered using
stdout however the conditional is not checking stdout and is attempting
to split a hash which fails rendering an obscure error. This change just
makes the conditional look at stdout and split on a sting.

Change-Id: I73c76fc83274e6f9a0157ad5a75797bebd682adb
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>
(cherry picked from commit 78ea059c93605f90019401bfc1c0ece3f5d2c242)