Correct failure when stopping cluster state is ignored
If the user passes in the `galera_ignore_cluster_state=true` flag, like
in the case of a node addition or a rolling replacement of nodes,
this role will fail to execute on all new nodes due to the existance of
a functioning cluster galera cluster. The first task in the pre-upgrade
steps attempts to stop the mysql service prior to running anything else.
This change makes it so the mysql service will not fail to stop on this
pre-upgrade task when the `galera_ignore_cluster_state=true` is passed
through.

Change-Id: Id9c4f11b3536e7f9d510f5ef8e6d82a82a96adac
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>