Correct failure when stopping cluster state is ignored
If the user passes in the `galera_ignore_cluster_state=true` flag, like
in the case of a node addition or a rolling replacement of nodes,
this role will fail to execute on all new nodes due to the existance of
a functioning cluster galera cluster. The first task in the pre-upgrade
steps attempts to stop the mysql service prior to running anything else.

In this patch we change the task flow to evaluate whether a service
exists, and if it does then to stop it - otherwise skip that task. This
ensures that if `galera_ignore_cluster_state=true` is passed through, the
playbook will run successfully.

Backport note:
This is a combined backport of https://review.openstack.org/607071 and
https://review.openstack.org/625929

Change-Id: Ib26c807057b9114e02ad32c759a5028555788363
(cherry picked from commit 69e2c736a07edc24ed1335bbf4d1a554a39fc40d)