Fix Galera self-signed SSL functionality
Ensure that `galera-req.pem` is removed when
`galera_ssl_self_signed_regen` is enabled. When this file is not
removed, the "Create galera ssl request" task does not run again, which
is required to create a new CSR and private key.

Copy the correct files to non-bootstrap Galera nodes, which are:
- CA certificate
- Private key for CSR
- Signed certificate

Rename a few variables for clarity between private key and certificate.

Change-Id: I3c65ff93498dde97e93fe9ac46ecee894a45c3e1