When upgrading, only stop the service if it exists
In Id9c4f11b3536e7f9d510f5ef8e6d82a82a96adac we implemented a
failed_when condition to ignore the state of the service stop task
when `galera_ignore_cluster_state=true` is passed through.

Unfortunately, the omit filter does not work in a task context [1],
resulting in the task always failing because the omit placeholder
string is evaluated as true. This makes the task always fail when
running an upgrade.

In this patch we change the task flow to evaluate whether a service
exists, and if it does then to stop it - otherwise skip that task.
This should cater for the rolling replacement use-case and will
work in every other situation too.

[1] https://github.com/ansible/ansible/issues/25678

Change-Id: Ib26c807057b9114e02ad32c759a5028555788363