Switch galera_install_dnf back to being a symlink
In I239a472c1d76f1cc00666280a93b73ecd84ba3d9 the galera_install_dnf
task file was mistakenly changed from a symlink to a copy of the
galera_install_yum file. Since then its contents have drifted from
the yum install process, but it's supposed to be exactly the same..

This patch switches it back to being a symlink to reduce the cost
of maintenance.

Change-Id: If7c40e7459e06526e883dc5a36b4fd31177598f1