Replace /etc/my.cnf.d if it exists for SUSE/CentOS
In SUSE/CentOS systems, /etc/my.cnf.d may already exist on the
system. We need to ensure that it is removed if that is true so
that we can replace it with symlinks to the directories expected
to be used by the MariaDB packages instead. By replacing it with
symlinks, we allow anything else on the system expecting those
paths to be in place to still work as before, but we also ensure
that the MariaDB package expectations are met.

Change-Id: I07fd0872a76af9dab89414207e297946e84965cf