Iterate over list of values of PPC packages dict
Currently as we are iterating over a higher level dictionary
and it is not possible to access the item.value.url attribute
directly.

Alternatively, we could get the values that is a list and
then iterate over that list to get properly the url and
checksum variables.

Also, we are not iterating over the dictionary we alternatively change
the item.key to galera_server_percona_distro_packages_alt_arch.keys()[0]
that produces the result as item.key.

Change-Id: I3e102377630d9c8ea5c9e3ea03cae3975a05dbb8
Closes-Bug: #1815902