[redhat] Update variables to be version agnostic
The RedHat-based operating systems such as CentOS have a pretty
stable list of packages, therefore, we don't need to pin it by
version and we can instead move to a much more generic redhat.yml
which will support a bigger range of systems.

This patch also sets up the mode for the /etc/mysql directory
to be 755 for things to work.  We should lock that down eventually
for only the MySQL user.

Change-Id: I2e9ee987a353a28da00cd37cbdb82ec46ccc414c