Use a private namespaced var for cluster name check
The galera cluster check uses "cluster_name" as a task return variable
when pulling the cluster name from an existing cluster. As one
might imagine, the cluster_name variable is very ambiguous and could
exist in a user's user_variables or inventory somewhere, which breaks
this check and fails the role execution. Instead of using 'cluster_name'
we can just return the task result to a private namespaced var
'_cluster_name' and use that for the assertion instead.

'node_status' is a similar check, also changed to '_node_status'

Change-Id: I0e530e67238a42e63f25ce0d59eefca2b7a2099a