Move to PyMySQL instead of MySQLDB for ansible-2.7
Running this module with ansible 2.7.13 to an Ubuntu 16.04 target was failing
with the following error message:

"The PyMySQL (Python 2.7 and Python 3.X) or MySQL-python (Python 2.X) module
is required."

Not sure why the fallback to using MysqlDB wasn't working, but it looks like
PyMySQL is the way forward.

This should fix https://bugs.launchpad.net/openstack-ansible/+bug/1824987.

Change-Id: Ic4ff3e24937d38e16fde843e350875c0024c92b2