Improve compatibility with connection delegation
In some cases when delegated hosts are used with connection plugins
and variables are used in the ansible_host var, they can be resolved
using the wrong node (the source of the delegated task, rather than
the target of the delegated task), causing the connection plugin to
run all of the mysql restarts on the same host, rather than delegating
to the correct ansible_host.

This minor fix improves compatibility in these situations so that
ansible_host is set to the correct host, causing the task to run
against the correct node.

Depends-On: https://review.opendev.org/687142
Change-Id: I082f4920cc8366e3b1309c952a5104d3d63215e5