Install xinetd clustercheck after mariadb is installed and setup
The configuration of mariadb is done mostly in handlers and previously
the code would install the xinetd clustercheck script before calling
flush_handlers to bootstrap and initialise the database. This leads
to a situation where the clustercheck may be attempting to interact
with the database during bootstrapping and before the monitoring user
has been created.

This change moves the setup tasks for the clustercheck service until
after the database has been bootstrapped and the users created. This
means that any healthchecks from the loadbalancer will not be seen by
the database until after the setup is completed during the initial
deployment.

Change-Id: Ibd55306a5c17bc0a4333e43725f1637420a2c9d9